requirements.txt
pytest
azure-identity
azure-mgmt-resource
